Create Multisender V0.1
Functions
constructor()
The constructor function is executed once when the contract is first deployed to the Ethereum blockchain. It sets the contract owner to the account that deploys the contract.

withdrawToken(address _token, uint256 _amount)
This function allows the contract owner to withdraw any ERC20 token that has been sent to the contract. It takes two parameters: the address of the ERC20 token to withdraw, and the amount of tokens to withdraw. The function transfers the specified amount of tokens to the contract owner's address.

multiSend(address[] memory _tokenAddresses, address[] memory _addresses, uint256[] memory _amounts, uint256 _nonce, string memory _password)
This function allows the user to send multiple ERC20 tokens to multiple addresses in one transaction. It takes four parameters:

_tokenAddresses: an array of ERC20 token addresses to send.
_addresses: an array of recipient addresses.
_amounts: an array of token amounts to send to each recipient.
_nonce: a unique identifier for the transaction to prevent duplicate transactions.
The function first checks that the length of each input array is equal, and that the transaction hasn't already been processed. Then, for each token in _tokenAddresses, the function transfers the corresponding amount from the sender's address to the corresponding recipient address using the ERC20 transferFrom function.

The function also requires the user to enter a password to confirm the transaction. The password is hardcoded as "password123" in this implementation.

usedHashes
This is a public mapping that keeps track of unique transaction IDs to prevent duplicate transactions. The mapping is keyed by a keccak256 hash of the _nonce parameter. Once a transaction has been processed, the corresponding hash is marked as used to prevent it from being used again in the future.